# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Etana Open Api Documentation",
        "description": "",
        "version": ""
    },
    "servers": [
        {
            "url": "https://utajeer.com/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/stripeToken": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create Stripe Token",
                "description": "Return Array",
                "operationId": "stripeToken",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "card_number": {
                                        "type": "string"
                                    },
                                    "exp_month": {
                                        "type": "string"
                                    },
                                    "exp_year": {
                                        "type": "string"
                                    },
                                    "cvc": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stripePost": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Payment create",
                "description": "return array",
                "operationId": "stripePost",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "stripeToken": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment Successfull.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/saveData": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Saved Data",
                "description": "Return string",
                "operationId": "saveData",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "payment_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data saved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}